refactor: simplify code structure
This is the result of watching too many talks on how to structure and
write idiomatic Go.
