+++ main.go
@@ -7,14 +7,119 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"text/tabwriter"
 )
 
-var (
-	cfg      = &Config{}
-	noCommas bool
-	noSpaces bool
-	showHelp bool
-)
+type format struct {
+	Name    string
+	Example string
+}
+
+var formats = [...]format{
+	{Name: "hex", Example: "#ebbcba"},
+	{Name: "hex-ns", Example: "ebbcba"},
+	{Name: "hsl", Example: "2, 55%, 83%"},
+	{Name: "hsl-array", Example: "[2, 55%, 83%]"},
+	{Name: "hsl-css", Example: "hsl(2deg 55% 83%)"},
+	{Name: "hsl-function", Example: "hsl(2, 55%, 83%)"},
+	{Name: "rgb", Example: "235, 188, 186"},
+	{Name: "rgb-ansi", Example: "235;188;186"},
+	{Name: "rgb-array", Example: "[235, 188, 186]"},
+	{Name: "rgb-function", Example: "rgb(235, 188, 186)"},
+}
+
+func formatsTable() string {
+	var sb strings.Builder
+	w := tabwriter.NewWriter(&sb, 1, 1, 1, ' ', 0)
+	for _, f := range formats {
+		fmt.Fprintf(w, "    %-13s %s\n", f.Name, f.Example)
+	}
+	w.Flush()
+	return sb.String()
+}
+
+func printFormatsTable() {
+	fmt.Fprint(os.Stdout, formatsTable())
+}
+
+func helpText() string {
+	return fmt.Sprintf(`
+  🌱 Bloom - The Rosé Pine theme generator
+
+  Usage
+    $ %s [options] <template>
+
+  Options
+    -o, --output <path>     Directory for generated files (default: dist)
+    -p, --prefix <string>   Color variable prefix (default: $)
+    -f, --format <format>   Color output format (default: hex)
+    -c, --create <variant>  Create template from existing theme (default: main)
+                            Variants: main, moon, dawn
+
+    --no-commas             Remove commas from color values
+    --no-spaces             Remove spaces from color values
+
+    -h, --help              Show help
+
+  Formats
+%s
+  Examples
+    $ %[1]s template.yaml
+    $ %[1]s --format hsl --output ./themes template.json
+    $ %[1]s --create dawn my-theme.toml
+
+`, os.Args[0], formatsTable())
+}
+
+func printHelp() {
+	fmt.Fprint(os.Stderr, helpText())
+}
+
+func main() {
+	cfg := &Config{}
+
+	flag.StringVar(&cfg.Output, "o", "dist", "")
+	flag.StringVar(&cfg.Output, "output", "dist", "")
+
+	flag.StringVar(&cfg.Create, "c", "", "")
+	flag.StringVar(&cfg.Create, "create", "", "")
+
+	flag.StringVar(&cfg.Prefix, "p", "$", "")
+	flag.StringVar(&cfg.Prefix, "prefix", "$", "")
+
+	flag.StringVar(&cfg.Format, "f", "hex", "")
+	flag.StringVar(&cfg.Format, "format", "hex", "")
+
+	noCommas := flag.Bool("no-commas", false, "")
+	noSpaces := flag.Bool("no-spaces", false, "")
+
+	flag.Usage = printHelp
+	flag.Parse()
+
+	args := flag.Args()
+
+	template, err := findTemplate(args)
+	if err != nil {
+		fmt.Fprintln(os.Stderr, "Error:", err)
+		printHelp()
+		os.Exit(1)
+	}
+
+	outputPassed := wasFlagPassed("o") || wasFlagPassed("output")
+	createPassed := wasFlagPassed("c") || wasFlagPassed("create")
+
+	if !outputPassed && createPassed {
+		cfg.Output = "."
+	}
+
+	cfg.Template = template
+	cfg.Commas = !*noCommas
+	cfg.Spaces = !*noSpaces
+
+	if err := Build(cfg); err != nil {
+		log.Fatal(err)
+	}
+}
 
 func wasFlagPassed(name string) bool {
 	for i := 1; i < len(os.Args); i++ {
@@ -22,14 +127,14 @@ func wasFlagPassed(name string) bool {
 		if arg == "-"+name || arg == "--"+name {
 			return true
 		}
-		if (arg == "-"+name || arg == "--"+name) && strings.Contains(arg, "=") {
+		if strings.Contains(arg, "=") && (strings.HasPrefix(arg, "-"+name+"=") || strings.HasPrefix(arg, "--"+name+"=")) {
 			return true
 		}
 	}
 	return false
 }
 
-func detectTemplate(args []string) (string, error) {
+func findTemplate(args []string) (string, error) {
 	switch len(args) {
 	case 0:
 		files, err := os.ReadDir(".")
@@ -47,66 +152,10 @@ func detectTemplate(args []string) (string, error) {
 				return name, nil
 			}
 		}
-		return "", fmt.Errorf("unable to find template")
-
+		return "", fmt.Errorf("unable to find template file")
 	case 1:
 		return args[0], nil
-
 	default:
 		return "", fmt.Errorf("multiple positional arguments detected, ensure all flags come before the template")
-	}
-}
-
-func init() {
-	flag.Usage = PrintHelp
-}
-
-func main() {
-	flag.StringVar(&cfg.Output, "o", "dist", "")
-	flag.StringVar(&cfg.Output, "output", "dist", "")
-
-	flag.StringVar(&cfg.Create, "c", "", "")
-	flag.StringVar(&cfg.Create, "create", "", "")
-
-	flag.StringVar(&cfg.Prefix, "p", "$", "")
-	flag.StringVar(&cfg.Prefix, "prefix", "$", "")
-
-	flag.StringVar(&cfg.Format, "f", "hex", "")
-	flag.StringVar(&cfg.Format, "format", "hex", "")
-
-	flag.BoolVar(&noCommas, "no-commas", false, "")
-	flag.BoolVar(&noSpaces, "no-spaces", false, "")
-
-	flag.BoolVar(&showHelp, "h", false, "")
-	flag.BoolVar(&showHelp, "help", false, "")
-
-	flag.Parse()
-
-	if showHelp {
-		PrintHelp()
-		return
-	}
-
-	args := flag.Args()
-
-	template, err := detectTemplate(args)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "Error:", err)
-		os.Exit(1)
-	}
-
-	outputPassed := wasFlagPassed("o") || wasFlagPassed("output")
-	createPassed := wasFlagPassed("c") || wasFlagPassed("create")
-
-	if !outputPassed && createPassed {
-		cfg.Output = "."
-	}
-
-	cfg.Template = template
-	cfg.Commas = !noCommas
-	cfg.Spaces = !noSpaces
-
-	if err := Build(cfg); err != nil {
-		log.Fatal(err)
 	}
 }
